Update map file to fit gender type changes in the schema

+++ core/model/modx/mysql/moduserprofile.map.inc.php
@@ -157,7 +157,7 @@ $xpdo_meta_map['modUserProfile']= array (
     ),
     'gender' => 
     array (
-      'dbtype' => 'int',
+      'dbtype' => 'tinyint',
       'precision' => '1',
       'phptype' => 'integer',
       'null' => false,
